# back/docs/医邻助手.openapi.json
{"openapi":"3.1.0","info":{"title":"医邻助手","description":"","version":"1.0.0"},"tags":[],"paths":{"/user/login":{"post":{"summary":"登录接口","deprecated":false,"description":"","tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"phone":"12345671901","password":"123456"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/user/queryPatientInformation":{"get":{"summary":"分页查询病人信息","deprecated":false,"description":"","tags":[],"parameters":[{"name":"page","in":"query","description":"","required":false,"example":"1","schema":{"type":"string"}},{"name":"size","in":"query","description":"","required":false,"example":"10","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/user/queryDoctorInformation":{"get":{"summary":"分页查询医生信息","deprecated":false,"description":"","tags":[],"parameters":[{"name":"page","in":"query","description":"","required":false,"example":"1","schema":{"type":"string"}},{"name":"size","in":"query","description":"","required":false,"example":"10","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/user/uploadUserMessage":{"post":{"summary":"批量导入账户信息","deprecated":false,"description":"","tags":[],"parameters":[],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"format":"binary","type":"string","example":""}},"required":["file"]}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/user/updatePatientInformation":{"put":{"summary":"更新患者个人信息","deprecated":false,"description":"","tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"uuid":"d59aa28a","nickname":"梨花","password":"1234567","home_addr":"xxx社区3栋2001号","phone":"19081309823"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/user/updatePhysicianInformation":{"put":{"summary":"更新医师个人信息","deprecated":false,"description":"","tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"uuid":"ea9b1129","nickname":"彰化","password":"1234567","hospital":"xxx社区医院","phone":"18923412345"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/user/adminLogin":{"post":{"summary":"管理员登录接口","deprecated":false,"description":"","tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"username":"admin","password":"admin"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/user/verifyToken":{"get":{"summary":"Token验证接口","deprecated":false,"description":"","tags":[],"parameters":[{"name":"Authorization","in":"header","description":"","required":false,"example":"","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/user/logOut":{"get":{"summary":"注销登录","deprecated":false,"description":"","tags":[],"parameters":[{"name":"Authorization","in":"header","description":"","required":false,"example":"","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/user/queryUserMessage":{"get":{"summary":"查询用户信息","deprecated":false,"description":"","tags":[],"parameters":[{"name":"Authorization","in":"header","description":"","required":true,"example":"","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/user/likeAccountByCV":{"get":{"summary":"分页按条件查询用户","deprecated":false,"description":"","tags":[],"parameters":[{"name":"cond","in":"query","description":"查询条件","required":false,"example":"username","schema":{"type":"string"}},{"name":"val","in":"query","description":"查询值","required":false,"example":"朱","schema":{"type":"string"}},{"name":"page","in":"query","description":"","required":false,"example":"1","schema":{"type":"string"}},{"name":"size","in":"query","description":"","required":false,"example":"10","schema":{"type":"string"}},{"name":"role","in":"query","description":"患者/医师","required":false,"example":"患者","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/inquiry/createInquiryRecord":{"post":{"summary":"创建问诊记录","deprecated":false,"description":"","tags":[],"parameters":[{"name":"Authorization","in":"header","description":"","required":true,"example":"","schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"type":"在家就医","inquiry_detail":"经常感到头痛","reserved_phone":"12345671901","appointment_time":"2024-10-02 20:30"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/inquiry/approveInquiryRecord":{"put":{"summary":"审批问诊记录","deprecated":false,"description":"","tags":[],"parameters":[{"name":"id","in":"query","description":"问诊登记ID","required":true,"example":"2","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/inquiry/appointedPhysician":{"post":{"summary":"指派医生","deprecated":false,"description":"","tags":[],"parameters":[{"name":"id","in":"query","description":"问诊登记ID","required":true,"example":"1","schema":{"type":"string"}},{"name":"userUUID","in":"query","description":"指派医师唯一标识符","required":true,"example":"ea9b1129","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/inquiry/queryAllInquiryRecord":{"get":{"summary":"分页查询所有问诊记录-管理员","deprecated":false,"description":"选题参数 isInquiry\n- 0: 默认值不需要填写\n- 1: 已经问诊结束\n- 2: 已经指派医师未问诊\n- 3: 暂未指派医师\n- 4: 指派医师代接诊\n- 5: 查询未审核的问诊记录","tags":[],"parameters":[{"name":"isInquiry","in":"query","description":"","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","description":"","required":false,"example":"1","schema":{"type":"string"}},{"name":"size","in":"query","description":"","required":false,"example":"10","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/inquiry/queryPhysicianInquiryRecord":{"get":{"summary":"分页查询医师所有问诊记录","deprecated":false,"description":"参数解析\n1. isInquiry 查询选项(默认为0，可不填)\n- 1 查询已经问诊结束的记录\n- 2 查询未问诊的记录\n- 4 查询未接诊的记录\n","tags":[],"parameters":[{"name":"isInquiry","in":"query","description":"","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","description":"","required":false,"example":"1","schema":{"type":"string"}},{"name":"size","in":"query","description":"","required":false,"example":"10","schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"","required":true,"example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3Mjk0NzgzODksInV1aWQiOiI0ODRlYjVmNyJ9.6LIpgP0o5bZ3_i4JV4kTQMZYv8LgpfbXOnEcIfM9gqM","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/inquiry/queryPatientInquiryRecord":{"get":{"summary":"分页查询患者所有问诊记录","deprecated":false,"description":"参数解析\n1. isInquriy 选项(默认为0，查询所有)\n- 1 查询所有已经审核通过记录\n- 2 待审核记录\n\n\n","tags":[],"parameters":[{"name":"isInquiry","in":"query","description":"","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","description":"","required":false,"example":"1","schema":{"type":"string"}},{"name":"size","in":"query","description":"","required":false,"example":"10","schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"","required":true,"example":"","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/inquiry/physicianReception":{"put":{"summary":"医师接诊","deprecated":false,"description":"","tags":[],"parameters":[{"name":"id","in":"query","description":"问诊登记ID","required":true,"example":"2","schema":{"type":"string"}},{"name":"isReception","in":"query","description":"是否接诊","required":true,"example":"false","schema":{"type":"boolean"}},{"name":"Authorization","in":"header","description":"接诊医生Token","required":true,"example":"","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/inquiry/queryInquiryRecordById":{"get":{"summary":"查询单个问诊记录","deprecated":false,"description":"","tags":[],"parameters":[{"name":"id","in":"query","description":"","required":true,"example":"1","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/inquiry/likeQueryInquiryRecord":{"get":{"summary":"分页按用户名条件查询病历","deprecated":false,"description":"","tags":[],"parameters":[{"name":"username","in":"query","description":"","required":false,"example":"朱","schema":{"type":"string"}},{"name":"isPatient","in":"query","description":"是否是患者 true查询患者，false查询医生","required":false,"example":"true","schema":{"type":"string"}},{"name":"page","in":"query","description":"","required":false,"example":"1","schema":{"type":"string"}},{"name":"size","in":"query","description":"","required":false,"example":"10","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/medical/queryMedialRecordInformation":{"get":{"summary":"查询病人病历信息","deprecated":false,"description":"","tags":[],"parameters":[{"name":"inquiryId","in":"query","description":"病历ID","required":true,"example":"10","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/medical/medicalRegistration":{"post":{"summary":"就诊登记","deprecated":false,"description":"1. diagnostic_description 病情描述(选填)\n2. bind_inquiry_id 绑定的问诊记录ID (必填)\n3. inquiry_video ipfs上传的问诊记录路由地址(必填)\n4. is_need_by_drug 是否需要代买药瓶 (必填)\n4. medical_img ipfs上传的病历图片路由地址 (必填)","tags":[],"parameters":[{"name":"Authorization","in":"header","description":"","required":true,"example":"","schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"diagnostic_description":"病人有高血压","bind_inquiry_id":2,"inquiry_video":"http://192.168.1.11:8080/ipfs/QmUSaebDyekM1GFk7FExm2NqMBFyTUC5crTuBTTSwiPAzG","medical_img":"http://192.168.1.11:8080/ipfs/QmQ64xb95NwKS3EpeUA9HEp3wYvi919FY936gJVt4mi9D5","is_need_by_drug":true}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/medical/queryMedicalByID":{"get":{"summary":"通过病历ID进行查询","deprecated":false,"description":"","tags":[],"parameters":[{"name":"medicalId","in":"query","description":"","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/drug/queryPatientAgentDrugHistory":{"get":{"summary":"分页患者查询药品代买历史记录","deprecated":false,"description":"## raw:\n- 0 默认查询所有数据\n- 1 查询医师已经购买的药品记录\n- 2 查询医师暂未购买的药品记录\n- 3 查询已经配送但是用户未收货的记录\n- 4 查询已经购买医师未派送的记录\n- 5 查询已经完成的历史记录","tags":[],"parameters":[{"name":"raw","in":"query","description":"","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","description":"","required":false,"example":"1","schema":{"type":"string"}},{"name":"size","in":"query","description":"","required":false,"example":"10","schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"","required":true,"example":"","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/drug/hospitalAgentDrugConfirmReceipt":{"post":{"summary":"医院管理人员审核药品代买情况","deprecated":false,"description":"","tags":[],"parameters":[{"name":"id","in":"query","description":"","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/drug/physiciansOrderAgentDrug":{"post":{"summary":"医师进行药品代卖","deprecated":false,"description":"","tags":[],"parameters":[{"name":"id","in":"query","description":"代买药品ID","required":true,"schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"","required":true,"example":"","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/drug/physiciansOrderDelivery":{"post":{"summary":"医师接单派送","deprecated":false,"description":"","tags":[],"parameters":[{"name":"id","in":"query","description":"代购ID","required":true,"example":"9","schema":{"type":"string"}},{"name":"certificate","in":"query","description":"上传凭证(购买药品发票以及配送到指定站点截图)","required":true,"example":"http://192.168.1.115:8080/ipfs/QmdgraqZUD48fKRRGK8ACGTZ7imhKrAQ3ah5bsGC9MvGd2","schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"","required":true,"example":"","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/drug/queryPhysiciansAgentHistoryRecord":{"get":{"summary":"分页医师查询代卖药品历史记录","deprecated":false,"description":"## raw:\n- 0 默认查询所有数据\n- 1 查询医师已经购买的药品记录\n- 2 查询医师暂未购买的药品记录\n- 3 查询已经配送但是用户未收货的记录\n- 4 查询已经购买医师未派送的记录\n- 5 查询已经完成的历史记录","tags":[],"parameters":[{"name":"raw","in":"query","description":"","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","description":"","required":false,"example":"1","schema":{"type":"string"}},{"name":"size","in":"query","description":"","required":false,"example":"10","schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"","required":true,"example":"","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/drug/queryDrugByID":{"get":{"summary":"输入代购ID查询","deprecated":false,"description":"","tags":[],"parameters":[{"name":"id","in":"query","description":"","required":true,"example":"9","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/drug/queryAllDrug":{"get":{"summary":"分页查询所有代购药品","deprecated":false,"description":"","tags":[],"parameters":[{"name":"page","in":"query","description":"","required":false,"example":"1","schema":{"type":"string"}},{"name":"size","in":"query","description":"","required":false,"example":"10","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/drug/queryDrugByMedicalId":{"get":{"summary":"输入病例ID查询相关药品信息","deprecated":false,"description":"","tags":[],"parameters":[{"name":"medical_id","in":"query","description":"病历ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/mock/generatePatientAccount":{"get":{"summary":"快速生成患者账户数据","deprecated":false,"description":"","tags":[],"parameters":[{"name":"number","in":"query","description":"测试账户数量","required":true,"example":"100","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/mock/generatePhysicianAccount":{"get":{"summary":"快速生成医师账户数据","deprecated":false,"description":"","tags":[],"parameters":[{"name":"number","in":"query","description":"","required":true,"example":"20","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/mock/testFullSystem":{"get":{"summary":"测试完整系统流程","deprecated":false,"description":"测试次数 = 测试次数 *  协程数量","tags":[],"parameters":[{"name":"number","in":"query","description":"测试次数","required":false,"example":"100","schema":{"type":"string"}},{"name":"goroutineNumber","in":"query","description":"启用协程数量","required":false,"example":"4","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/sum/getInquirySumTypeDay":{"get":{"summary":"输入最近天数获取问诊类型每天数量统计","deprecated":false,"description":"","tags":[],"parameters":[{"name":"day","in":"query","description":"获取最近几天数据","required":false,"example":"7","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/sum/getDrugSumTypeDay":{"get":{"summary":"输入最近天数获取药品订单每日统计","deprecated":false,"description":"","tags":[],"parameters":[{"name":"day","in":"query","description":"","required":false,"example":"7","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/v1/chat/completions":{"post":{"summary":"glm4","deprecated":false,"description":"","tags":[],"parameters":[{"name":"Authorization","in":"header","description":"","required":false,"example":"Bearer sk-fastgptnb","schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"model":"glm4","messages":[{"role":"system","content":"你是一个经验丰富的AI助手，能帮我处理很多计算机事务"},{"role":"user","content":"你好 1+1=?"},{"role":"assistant","content":""}]}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/ai/getMedicalSum":{"get":{"summary":"获取病历信息总结","deprecated":false,"description":"","tags":[],"parameters":[{"name":"Authorization","in":"header","description":"","required":true,"example":"","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/upload":{"post":{"summary":"上传文件至ipfs接口","deprecated":false,"description":"","tags":[],"parameters":[],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"example":"","type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}}},"components":{"schemas":{},"securitySchemes":{}},"servers":[{"url":"http://192.168.1.108:8088/api","description":"开发环境"}]}